// src/apis/generated/models/addressInfo.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Project Name
 * 개발: [api.dev.my-domain.com](https://api.dev.my-domain.com)
<br/>
운영: [api.my-domain.com](https://api.my-domain.com)

 * OpenAPI spec version: 1.0.0
 */

/**
 * @nullable
 */
export type AddressInfo = {
  country?: string
  city?: string
  address?: string
  zipcode?: string
} | null