// src/apis/generated/models/registerAdditionalInfo.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Project Name
 * 개발: [api.dev.my-domain.com](https://api.dev.my-domain.com)
<br/>
운영: [api.my-domain.com](https://api.my-domain.com)

 * OpenAPI spec version: 1.0.0
 */
import type { AddressInfo } from './addressInfo'
import type { RegisterAdditionalInfoGender } from './registerAdditionalInfoGender'
import type { SnsInfo } from './snsInfo'

export interface RegisterAdditionalInfo {
  nickname: string
  /** @nullable */
  gender?: RegisterAdditionalInfoGender
  isReceiveSeedingBox: boolean
  addressInfo?: AddressInfo
  snsInfo?: SnsInfo
}