// src/apis/generated/models/registerUserRequestDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Project Name
 * 개발: [api.dev.my-domain.com](https://api.dev.my-domain.com)
<br/>
운영: [api.my-domain.com](https://api.my-domain.com)

 * OpenAPI spec version: 1.0.0
 */
import type { RegisterAdditionalInfo } from './registerAdditionalInfo'

export interface RegisterUserRequestDto {
  /** 이름 */
  firstName: string
  lastName: string
  birthDate: string
  email: string
  /** email verifier를 통해 얻은 token값 입니다. */
  emailToken: string
  password: string
  passwordConfirm: string
  phoneCountryCode: string
  phoneNumber: string
  isTermsAgreed: boolean
  isPrivacyAgreed: boolean
  additionalInfo?: RegisterAdditionalInfo
}