// src/apis/generated/models/snsInfo.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Project Name
 * 개발: [api.dev.my-domain.com](https://api.dev.my-domain.com)
<br/>
운영: [api.my-domain.com](https://api.my-domain.com)

 * OpenAPI spec version: 1.0.0
 */
import type { SnsInfoPlatform } from './snsInfoPlatform'

/**
 * @nullable
 */
export type SnsInfo = {
  platform?: SnsInfoPlatform
  snsId?: string
  /** platform이 others인 경우 sns 풀랫폼 명 */
  otherPlatformName?: string
} | null